Improve description of eChronos as a family of RTOSes
eChronos has always been about generating a specific RTOS for a specific
use-case / hardware combination. We've struggled to describe that well
in the documentation previously.

This change of calling eChronos a "family" of RTOS-s and each specific
configuration an actual RTOS is hopefully a bit of an improvement and
helps newcomers understand how best to use eChronos.