Exclude some files from having the standard header added during release
It is not appropriate for all files in the release to have the standard
CSIRO header attached automatically. This change allows for specified
files to be explicitly excluded from this regime, instead relying on the
file itself to have any necessary copyright and license information
directly embedded in the file.