Add ARMv8A startup support for multiple cores
A new cpu_count configuration point is added to armv8a.startup module.
This has a default of 1 (so existing code should work without any changes).

The linker script is updated to create multiple stacks (one for each
configured CPU).

The startup assembly is updated to correctly initialise the stack for
each configured CPU. Core 0 is used to initialize BSS; other CPUs
jump directly to platform_init / main.

It is up to platform code to coordinate CPUs appropriately (i.e.
so that secondary cores don't execute code that relies on the BSS
until after the BSS is initialised).