Fix bug with schema default empty lists
A feature of the schema language is that it is possible to have a
default empty list.

The existing implementation has a bug such that the same empty list
instance is returned each time the default is used. The updated
implementation creates a new empty list each time the default is
required.